pyproject.toml
[project]
name = "npci-rss-feeds"
version = "0.1.0"
description = "Generate RSS Feeds for NPCI products"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "curl-cffi>=0.13.0",
    "feedgen>=1.0.0",
    "pyyaml>=6.0.2",
]
